requirements.txt
black==23.7.0
cfgv==3.3.1
click==8.1.5
distlib==0.3.6
fastcore==1.5.29
filelock==3.12.2
ghapi==1.0.4
identify==2.5.24
isort==5.12.0
libcst==0.4.10
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.0
pre-commit==3.3.3
pycln==2.1.5
python-dotenv==1.0.0
PyYAML==6.0
tomlkit==0.11.8
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
virtualenv==20.24.0
